Add files via upload
This is the first version of movie booking website, here you can Create, Delete movies from manage movies tab on navigation bar. Add Audi and timings for screening of the movie in the theater. Also map the movies database to the screening database. You can view all the screening movies on the homepage.